moved NewSmurf form up one level
